requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langsmith==0.1.93
Levenshtein==0.25.1
regex==2024.7.24
reportlab==4.2.2
selenium==4.9.1
webdriver-manager==4.0.2
click
requests==2.31.0
google-auth 
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pyyaml
langchain
git+https://github.com/feder-cr/lib_resume_builder_AIHawk.git

